♻️ refactor: Order 클래스 검증 로직 분리
